Problem: Tendermint RPC client does not support websocket over TLS
Solution: Moved to `tokio-tungstenite` with support for websockets over TLS. Fixes #1346.
