+++ client-common/src/tendermint.rs
@@ -1,14 +1,14 @@
 //! Tendermint client operations
 mod client;
-mod unauthorized_client;
 #[cfg(feature = "websocket-rpc")]
-mod websocket_rpc_client;
+mod rpc_client;
+mod unauthorized_client;
 
 pub mod lite;
 pub mod mock;
 pub mod types;
 
 pub use client::Client;
-pub use unauthorized_client::UnauthorizedClient;
 #[cfg(feature = "websocket-rpc")]
-pub use websocket_rpc_client::WebsocketRpcClient;
+pub use rpc_client::WebsocketRpcClient;
+pub use unauthorized_client::UnauthorizedClient;
