+++ client-common/src/tendermint/rpc_client.rs
@@ -1,0 +1,7 @@
+mod async_rpc_client;
+mod sync_rpc_client;
+mod types;
+mod websocket_rpc_loop;
+
+pub use async_rpc_client::AsyncRpcClient;
+pub use sync_rpc_client::SyncRpcClient as WebsocketRpcClient;
