--- client-common/src/tendermint/websocket_rpc_client.rs
@@ -1,465 +1,0 @@
-mod types;
-mod websocket_rpc_loop;
-
-pub use types::ConnectionState;
-
-use itertools::izip;
-use std::collections::HashMap;
-use std::net::TcpStream;
-use std::sync::mpsc::{sync_channel, Receiver, SyncSender};
-use std::sync::{atomic::AtomicUsize, atomic::Ordering, Arc, Mutex};
-use std::thread;
-use std::time::{Duration, SystemTime};
-
-use serde::Deserialize;
-use serde_json::{json, Value};
-use tendermint::{lite, validator};
-use tungstenite::{Message, WebSocket};
-
-use self::types::*;
-use crate::tendermint::types::*;
-use crate::tendermint::{lite::TrustedState, Client};
-use crate::{Error, ErrorKind, Result, ResultExt};
-use chain_core::state::ChainState;
-
-const RESPONSE_TIMEOUT: Duration = Duration::from_secs(10);
-
-const WAIT_FOR_CONNECTION_SLEEP_INTERVAL: Duration = Duration::from_millis(200);
-const WAIT_FOR_CONNECTION_COUNT: usize = 50;
-
-/// Tendermint RPC Client (uses websocket in transport layer)
-#[derive(Clone)]
-pub struct WebsocketRpcClient {
-    connection_state: Arc<Mutex<ConnectionState>>,
-    websocket_writer: Arc<Mutex<WebSocket<TcpStream>>>,
-    channel_map: Arc<Mutex<HashMap<String, SyncSender<JsonRpcResponse>>>>,
-    unique_id: Arc<AtomicUsize>,
-}
-
-impl WebsocketRpcClient {
-    /// Creates a new instance of `WebsocketRpcClient`
-    //
-    // # How it works
-    //
-    // - Spawns `websocket_rpc_loop`.
-    // - Spawns `websocket_rpc_loop` monitor.
-    pub fn new(url: &str) -> Result<Self> {
-        let channel_map: Arc<Mutex<HashMap<String, SyncSender<JsonRpcResponse>>>> =
-            Default::default();
-
-        let (websocket_reader, websocket_writer) = websocket_rpc_loop::new_connection(url)?;
-        let websocket_writer = Arc::new(Mutex::new(websocket_writer));
-
-        let loop_handle = websocket_rpc_loop::spawn(
-            channel_map.clone(),
-            websocket_reader,
-            websocket_writer.clone(),
-        );
-
-        let connection_state = websocket_rpc_loop::monitor(
-            url.to_owned(),
-            channel_map.clone(),
-            loop_handle,
-            websocket_writer.clone(),
-        );
-
-        Ok(Self {
-            connection_state,
-            websocket_writer,
-            channel_map,
-            unique_id: Arc::new(AtomicUsize::new(0)),
-        })
-    }
-
-    /// Returns current connection state of websocket connection
-    pub fn connection_state(&self) -> ConnectionState {
-        *self
-            .connection_state
-            .lock()
-            .expect("Unable to acquire lock on connection state")
-    }
-
-    /// Sends a RPC request
-    //
-    // # How it works
-    //
-    // - Prepares JSON-RPC request from `method` and `params` (generates a random `request_id`).
-    // - Creates a `sync_channel` pair
-    // - Inserts `channel_sender` to `channel_map` corresponding to generated `request_id`.
-    // - Ensure that the `websocket_rpc_loop` is in `Connected` state.
-    // - Send request websocket message.
-    // - Receive response on `channel_receiver`.
-    fn request(&self, method: &str, params: &[Value]) -> Result<Value> {
-        let (id, channel_receiver) = self.send_request(method, params)?;
-        self.receive_response(method, params, &id, channel_receiver)
-    }
-
-    /// Sends RPC requests for a batch.
-    ///
-    /// # Note
-    ///
-    /// This does not use batch JSON-RPC requests but makes multiple single JSON-RPC requests in parallel.
-    fn request_batch(&self, batch_params: &[(&str, Vec<Value>)]) -> Result<Vec<Value>> {
-        let mut receivers = Vec::with_capacity(batch_params.len());
-
-        for (method, params) in batch_params.iter() {
-            let (id, channel_receiver) = self.send_request(method, &params)?;
-            receivers.push((id, channel_receiver));
-        }
-
-        receivers
-            .into_iter()
-            .zip(batch_params.iter().cloned()) // FIXME fix the clone
-            .map(|((id, channel_receiver), (method, params))| {
-                self.receive_response(&method, &params, &id, channel_receiver)
-            })
-            .collect()
-    }
-
-    /// Sends a JSON-RPC request and returns `request_id` and `response_channel`
-    fn send_request(
-        &self,
-        method: &str,
-        params: &[Value],
-    ) -> Result<(String, Receiver<JsonRpcResponse>)> {
-        let id = self.unique_id.fetch_add(1, Ordering::Relaxed).to_string();
-        let message = prepare_message(&id, method, params)?;
-        let (channel_sender, channel_receiver) = sync_channel::<JsonRpcResponse>(1);
-
-        self.channel_map
-            .lock()
-            .expect("Unable to acquire lock on websocket request map: Lock is poisoned")
-            .insert(id.clone(), channel_sender);
-
-        self.ensure_connected()?;
-
-        self.websocket_writer
-            .lock()
-            .expect("Unable to acquire lock on websocket writer: Lock is poisoned")
-            .write_message(message)
-            .chain(|| {
-                (
-                    ErrorKind::InternalError,
-                    "Unable to send message to websocket writer",
-                )
-            })
-            .map_err(|err| {
-                self.channel_map
-                    .lock()
-                    .expect("Unable to acquire lock on websocket request map: Lock is poisoned")
-                    .remove(&id);
-                err
-            })?;
-
-        Ok((id, channel_receiver))
-    }
-
-    /// Receives response from websocket for given id.
-    fn receive_response(
-        &self,
-        method: &str,
-        params: &[Value],
-        id: &str,
-        receiver: Receiver<JsonRpcResponse>,
-    ) -> Result<Value> {
-        let response = receiver
-            .recv_timeout(RESPONSE_TIMEOUT)
-            .chain(|| {
-                (
-                    ErrorKind::InternalError,
-                    "Unable to receive message from channel receiver",
-                )
-            })
-            .map_err(|err| {
-                self.channel_map
-                    .lock()
-                    .expect("Unable to acquire lock on websocket request map: Lock is poisoned")
-                    .remove(id);
-                err
-            })?;
-
-        if let Some(err) = response.error {
-            Err(Error::new_with_source(
-                ErrorKind::TendermintRpcError,
-                format!(
-                    "Error response from tendermint RPC for request method ({}) and params ({:?})",
-                    method, params
-                ),
-                Box::new(err),
-            ))
-        } else {
-            Ok(response.result.unwrap_or_default())
-        }
-    }
-
-    /// Ensures that the websocket is connected.
-    fn ensure_connected(&self) -> Result<()> {
-        for _ in 0..WAIT_FOR_CONNECTION_COUNT {
-            if ConnectionState::Connected
-                == *self
-                    .connection_state
-                    .lock()
-                    .expect("Unable to acquire lock on connection state")
-            {
-                return Ok(());
-            }
-
-            thread::sleep(WAIT_FOR_CONNECTION_SLEEP_INTERVAL);
-        }
-
-        Err(Error::new(
-            ErrorKind::InternalError,
-            "Websocket connection disconnected",
-        ))
-    }
-
-    /// Makes an RPC call and deserializes response
-    pub fn call<T>(&self, method: &str, params: &[Value]) -> Result<T>
-    where
-        for<'de> T: Deserialize<'de>,
-    {
-        let response_value = self.request(method, params)?;
-        serde_json::from_value(response_value).chain(|| {
-            (
-                ErrorKind::DeserializationError,
-                format!(
-                    "Unable to deserialize `{}` from JSON-RPC response for params: {:?}",
-                    method, params
-                ),
-            )
-        })
-    }
-
-    /// Makes RPC call in batch and deserializes responses
-    pub fn call_batch<T>(&self, params: &[(&str, Vec<Value>)]) -> Result<Vec<T>>
-    where
-        for<'de> T: Deserialize<'de>,
-    {
-        if params.is_empty() {
-            // Do not send empty batch requests
-            return Ok(Default::default());
-        }
-
-        if params.len() == 1 {
-            // Do not send batch request when there is only one set of params
-            self.call::<T>(params[0].0, &params[0].1)
-                .map(|value| vec![value])
-        } else {
-            let response_values = self.request_batch(params)?;
-
-            response_values
-                .into_iter()
-                .zip(params.iter().cloned()) // FIXME fix the clone
-                .map(|(response_value, (method, params))| {
-                    serde_json::from_value(response_value).chain(|| {
-                        (
-                            ErrorKind::DeserializationError,
-                            format!(
-                                "Unable to deserialize `{}` from JSON-RPC response for params: {:?}",
-                                method, params
-                            ),
-                        )
-                    })
-                })
-                .collect()
-        }
-    }
-
-    fn validators_batch<T: Iterator<Item = u64>>(
-        &self,
-        heights: T,
-    ) -> Result<Vec<ValidatorsResponse>> {
-        let params = heights
-            .map(|height| ("validators", vec![json!(height.to_string())]))
-            .collect::<Vec<(&str, Vec<Value>)>>();
-        self.call_batch(&params)
-    }
-
-    fn commit_batch<'a, T: Iterator<Item = &'a u64>>(
-        &self,
-        heights: T,
-    ) -> Result<Vec<CommitResponse>> {
-        let params = heights
-            .map(|height| ("commit", vec![json!(height.to_string())]))
-            .collect::<Vec<(&str, Vec<Value>)>>();
-        self.call_batch(&params)
-    }
-}
-
-impl Client for WebsocketRpcClient {
-    #[inline]
-    fn genesis(&self) -> Result<Genesis> {
-        Ok(self
-            .call::<GenesisResponse>("genesis", Default::default())?
-            .genesis)
-    }
-
-    #[inline]
-    fn status(&self) -> Result<Status> {
-        self.call("status", Default::default())
-    }
-
-    #[inline]
-    fn block(&self, height: u64) -> Result<Block> {
-        let params = [json!(height.to_string())];
-        Ok(self.call::<BlockResponse>("block", &params)?.block)
-    }
-
-    #[inline]
-    fn block_batch<'a, T: Iterator<Item = &'a u64>>(&self, heights: T) -> Result<Vec<Block>> {
-        let params = heights
-            .map(|height| ("block", vec![json!(height.to_string())]))
-            .collect::<Vec<(&str, Vec<Value>)>>();
-        let rsps = self.call_batch::<BlockResponse>(&params)?;
-        Ok(rsps.into_iter().map(|rsp| rsp.block).collect())
-    }
-
-    #[inline]
-    fn block_results(&self, height: u64) -> Result<BlockResults> {
-        let params = [json!(height.to_string())];
-        self.call("block_results", &params)
-    }
-
-    #[inline]
-    fn block_results_batch<'a, T: Iterator<Item = &'a u64>>(
-        &self,
-        heights: T,
-    ) -> Result<Vec<BlockResults>> {
-        let params = heights
-            .map(|height| ("block_results", vec![json!(height.to_string())]))
-            .collect::<Vec<(&str, Vec<Value>)>>();
-        self.call_batch(&params)
-    }
-
-    fn broadcast_transaction(&self, transaction: &[u8]) -> Result<BroadcastTxResponse> {
-        let params = [json!(transaction)];
-        let rsp = self.call::<BroadcastTxResponse>("broadcast_tx_sync", &params)?;
-
-        if rsp.code.is_err() {
-            Err(Error::new(ErrorKind::TendermintRpcError, rsp.log.as_ref()))
-        } else {
-            Ok(rsp)
-        }
-    }
-
-    fn query(&self, path: &str, data: &[u8]) -> Result<AbciQuery> {
-        let params = [
-            json!(path),
-            json!(hex::encode(data)),
-            json!(null),
-            json!(null),
-        ];
-        let result = self
-            .call::<AbciQueryResponse>("abci_query", &params)?
-            .response;
-
-        if result.code.is_err() {
-            return Err(Error::new(
-                ErrorKind::TendermintRpcError,
-                result.log.to_string(),
-            ));
-        }
-
-        Ok(result)
-    }
-
-    fn block_batch_verified<'a, T: Clone + Iterator<Item = &'a u64>>(
-        &self,
-        mut state: TrustedState,
-        heights: T,
-    ) -> Result<(Vec<Block>, TrustedState)> {
-        let commits = self.commit_batch(heights.clone())?;
-        let validators: Vec<validator::Set> = self
-            .validators_batch(heights.clone().map(|h| h.saturating_add(1)))?
-            .into_iter()
-            .map(|rsp| validator::Set::new(rsp.validators))
-            .collect();
-        let blocks = self.block_batch(heights)?;
-        for (commit, next_vals, block) in izip!(&commits, &validators, &blocks) {
-            let signed_header =
-                lite::SignedHeader::new(commit.signed_header.clone(), block.header.clone());
-            state = if let Some(state) = &state.0 {
-                lite::verifier::verify_single(
-                    state.clone(),
-                    &signed_header,
-                    state.validators(),
-                    next_vals,
-                    // FIXME make parameters configurable
-                    lite::TrustThresholdFraction::new(1, 3).unwrap(),
-                    Duration::from_secs(std::u32::MAX as u64),
-                    SystemTime::now(),
-                )
-                .map_err(|err| {
-                    Error::new(
-                        ErrorKind::VerifyError,
-                        format!("block verify failed: {:?}", err),
-                    )
-                })?
-                .into()
-            } else {
-                // TODO verify block1 against genesis block
-                lite::TrustedState::new(signed_header, next_vals.clone()).into()
-            };
-        }
-        Ok((blocks, state))
-    }
-
-    fn query_state_batch<T: Iterator<Item = u64>>(&self, heights: T) -> Result<Vec<ChainState>> {
-        let params: Vec<(&str, Vec<Value>)> = heights
-            .map(|height| {
-                (
-                    "abci_query",
-                    vec![
-                        json!("state"),
-                        json!(null),
-                        json!(height.to_string()),
-                        json!(null),
-                    ],
-                )
-            })
-            .collect();
-        let rsps = self.call_batch::<AbciQueryResponse>(&params)?;
-
-        rsps.into_iter()
-            .map(|rsp| {
-                if let Some(value) = rsp.response.value {
-                    let state = serde_json::from_str(
-                        &String::from_utf8(value)
-                            .chain(|| (ErrorKind::InvalidInput, "chain state decode failed"))?,
-                    )
-                    .chain(|| (ErrorKind::InvalidInput, "chain state decode failed"))?;
-                    Ok(state)
-                } else {
-                    Err(Error::new(
-                        ErrorKind::InvalidInput,
-                        format!(
-                            "abci query fail: {}, {}",
-                            rsp.response.code.value(),
-                            rsp.response.log,
-                        ),
-                    ))
-                }
-            })
-            .collect()
-    }
-}
-
-fn prepare_message(id: &str, method: &str, params: &[Value]) -> Result<Message> {
-    let request = JsonRpcRequest {
-        id,
-        jsonrpc: "2.0",
-        method,
-        params,
-    };
-
-    let request_json = serde_json::to_string(&request).chain(|| {
-        (
-            ErrorKind::SerializationError,
-            "Unable to serialize RPC request to json",
-        )
-    })?;
-
-    let message = Message::Text(request_json);
-
-    Ok(message)
-}
